fix: daily email routine crashes app
