+++ cmd/mccs-alpha-api/main.go
@@ -26,16 +26,19 @@ func main() {
 // ServeBackGround performs the background activities.
 func ServeBackGround() {
 	c := cron.New()
+
 	viper.SetDefault("daily_email_schedule", "0 0 7 * * *")
 	c.AddFunc(viper.GetString("daily_email_schedule"), func() {
 		l.Logger.Info("[ServeBackGround] Running daily email schedule. \n")
 		dailyemail.Run()
 	})
+
 	viper.SetDefault("balance_check_schedule", "0 0 * * * *")
 	c.AddFunc(viper.GetString("balance_check_schedule"), func() {
 		l.Logger.Info("[ServeBackGround] Running balance check schedule. \n")
 		balancecheck.Run()
 	})
+
 	c.Start()
 }
