+++ internal/app/repository/es/tag.go
@@ -158,8 +158,6 @@ func (es *tag) DeleteByID(id string) error {
 	return nil
 }
 
-// TO BE REMOVED
-
 // MatchOffer matches wants for the given offer.
 func (es *tag) MatchOffer(offer string, lastLoginDate time.Time) ([]string, error) {
 	q := newTagQuery(offer, lastLoginDate, "wantAddedAt")
@@ -172,7 +170,7 @@ func (es *tag) MatchOffer(offer string, lastLoginDate time.Time) ([]string, erro
 		return nil, err
 	}
 
-	matchTags := make([]string, 0, 8)
+	matchTags := []string{}
 	for _, hit := range res.Hits.Hits {
 		var record types.TagESRecord
 		err := json.Unmarshal(hit.Source, &record)
@@ -197,7 +195,7 @@ func (es *tag) MatchWant(want string, lastLoginDate time.Time) ([]string, error)
 		return nil, err
 	}
 
-	matchTags := make([]string, 0, 8)
+	matchTags := []string{}
 	for _, hit := range res.Hits.Hits {
 		var record types.TagESRecord
 		err := json.Unmarshal(hit.Source, &record)
